Move UserTypeDataNode to the correct namespace
