+++ SAGESharp.Tests/IO/Binary/TreeBasedSerialization/UserTypeDataNodeTests.cs
@@ -9,10 +9,11 @@ using NSubstitute.ClearExtensions;
 using NUnit.Framework;
 using SAGESharp.IO;
 using SAGESharp.IO.Binary;
+using SAGESharp.IO.Binary.TreeBasedSerialization;
 using System;
 using System.Collections.Generic;
 
-namespace SAGESharp.Tests.IO
+namespace SAGESharp.Tests.IO.Binary.TreeBasedSerialization
 {
     class UserTypeDataNodeTests
     {
