+++ SAGESharp/IO/TreeBasedSerialization.cs
@@ -387,34 +387,5 @@ namespace SAGESharp.IO
 
         private static bool IsString(object value) => typeof(string).Equals(value.GetType());
     }
-
-    internal sealed class UserTypeDataNode<T> : IDataNode where T : new()
-    {
-        public UserTypeDataNode(IReadOnlyList<IEdge> edges)
-        {
-            Validate.ArgumentNotNull(edges, nameof(edges));
-            Validate.Argument(edges != null, $"{nameof(edges)} should not be empty");
-
-            Edges = edges;
-        }
-
-        public IReadOnlyList<IEdge> Edges { get; }
-
-        public object Read(IBinaryReader binaryReader)
-        {
-            Validate.ArgumentNotNull(binaryReader, nameof(binaryReader));
-
-            return new T();
-        }
-
-        public void Write(IBinaryWriter binaryWriter, object value)
-        {
-            Validate.ArgumentNotNull(binaryWriter, nameof(binaryWriter));
-            Validate.ArgumentNotNull(value, nameof(value));
-            Validate.Argument(IsType(value), $"Cannot write value of type {value.GetType().Name} as type {typeof(T).Name}.");
-        }
-
-        private static bool IsType(object value) => typeof(T) == value.GetType();
-    }
     #endregion
 }
