+++ SAGESharp/SAGESharp.csproj
@@ -105,6 +105,7 @@
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeBuilder.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeReader.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeWriter.cs" />
+    <Compile Include="IO\Binary\TreeBasedSerialization\UserTypeDataNode.cs" />
     <Compile Include="IO\Binary\Writer.cs" />
     <Compile Include="IO\SerializablePropertyAttribute.cs" />
     <Compile Include="IO\TreeBasedSerialization.cs" />
